+++ src/css/datagrid.css
@@ -1,1 +1,1 @@
-.dg-wrapper{display:grid;grid-auto-flow:row;width:99vw}.dg-wrapper *{color:#fff}.dg-wrapper .dg-header{display:grid;grid-template-columns:6fr 2fr 2fr 1fr 1fr 1fr;gap:4px;background:rgba(32,82,149,.25);border:1px solid #205295;border-radius:16px;margin-bottom:12px;box-shadow:0px 0px 5px rgba(0,0,0,.5)}.dg-wrapper .dg-body{display:grid;grid-auto-flow:row;gap:8px}.dg-wrapper .dg-body .dg-row{display:grid;grid-template-rows:auto 1fr;background:rgba(20,66,114,.25);border:1px solid #144272;border-radius:16px;box-shadow:0px 0px 5px rgba(0,0,0,.5);transition:.05s ease-in-out}.dg-wrapper .dg-body .dg-row .dg-row-master{display:grid;grid-template-columns:6fr 2fr 2fr 1fr 1fr 1fr;gap:4px;align-items:center}.dg-wrapper .dg-body .dg-row:hover{cursor:pointer;filter:brightness(1.2)}.dg-wrapper .dg-header-cell,.dg-wrapper .dg-cell{padding:6px 10px}.dg-wrapper .row-selected .dg-row-master{border-bottom:1px solid #144272}.dg-wrapper .dg-center{text-align:center}.dtg-wrapper{display:grid;grid-auto-flow:row;padding:16px}.dtg-wrapper .dtg-header{display:grid;grid-template-columns:1fr 1fr 1fr 2fr 1fr;gap:4px;background:rgba(44,116,179,.25);border:1px solid #2c74b3;border-radius:16px;margin-bottom:12px;box-shadow:0px 0px 5px rgba(0,0,0,.5)}.dtg-wrapper .dtg-body{display:grid;grid-auto-flow:row;gap:6px}.dtg-wrapper .dtg-body .dtg-row{display:grid;grid-template-columns:1fr 1fr 1fr 2fr 1fr;background:rgba(32,82,149,.25);border:1px solid #205295;border-radius:16px;box-shadow:0px 0px 5px rgba(0,0,0,.5);gap:4px;align-items:center}.dtg-wrapper .dtg-header-cell,.dtg-wrapper .dtg-cell{padding:6px 8px}.loading-icon{justify-self:center;margin:8px;width:50px;height:50px;border:5px solid rgba(255,255,255,.3);border-top:5px solid #fff;border-radius:50%;animation:spin 1s linear infinite}@keyframes spin{from{transform:rotate(0deg)}to{transform:rotate(360deg)}}.mobile-warning{border:1px solid red;border-radius:16px;background:rgba(255,0,0,.25);padding:8px 12px;text-align:center;font-weight:900}@media(min-width: 1280px){.mobile-warning{display:none}}/*# sourceMappingURL=datagrid.css.map */
+.dg-wrapper{display:grid;grid-auto-flow:row;width:99vw}.dg-wrapper *{color:#fff}.dg-wrapper .dg-header{display:grid;grid-template-columns:6fr 2fr 2fr 1fr 1fr 1fr;gap:4px;background:rgba(32,82,149,.25);border:1px solid #205295;border-radius:16px;margin-bottom:12px;box-shadow:0px 0px 5px rgba(0,0,0,.5)}.dg-wrapper .dg-body{display:grid;grid-auto-flow:row;gap:8px}.dg-wrapper .dg-body .dg-row{display:grid;grid-template-rows:auto 1fr;background:rgba(20,66,114,.25);border:1px solid #144272;border-radius:16px;box-shadow:0px 0px 5px rgba(0,0,0,.5);transition:.05s ease-in-out}.dg-wrapper .dg-body .dg-row .dg-row-master{display:grid;grid-template-columns:6fr 2fr 2fr 1fr 1fr 1fr;gap:4px;align-items:center}.dg-wrapper .dg-body .dg-row:hover{cursor:pointer;filter:brightness(1.2)}.dg-wrapper .dg-header-cell,.dg-wrapper .dg-cell{padding:6px 10px}.dg-wrapper .row-selected .dg-row-master{border-bottom:1px solid #144272}.dg-wrapper .dg-center{text-align:center}.dtg-wrapper{display:grid;grid-auto-flow:row;padding:16px}.dtg-wrapper .dtg-header{display:grid;grid-template-columns:1fr 1fr 1fr 2fr 1fr;gap:4px;background:rgba(44,116,179,.25);border:1px solid #2c74b3;border-radius:16px;margin-bottom:8px;box-shadow:0px 0px 5px rgba(0,0,0,.5)}.dtg-wrapper .dtg-body{display:grid;grid-auto-flow:row;gap:4px}.dtg-wrapper .dtg-body .dtg-row{display:grid;grid-template-columns:1fr 1fr 1fr 2fr 1fr;background:rgba(32,82,149,.25);border:1px solid #205295;border-radius:16px;box-shadow:0px 0px 5px rgba(0,0,0,.5);gap:4px;align-items:center}.dtg-wrapper .dtg-header-cell,.dtg-wrapper .dtg-cell{padding:6px 8px}.loading-icon{justify-self:center;margin:8px;width:50px;height:50px;border:5px solid rgba(255,255,255,.3);border-top:5px solid #fff;border-radius:50%;animation:spin 1s linear infinite}@keyframes spin{from{transform:rotate(0deg)}to{transform:rotate(360deg)}}.mobile-warning{border:1px solid red;border-radius:16px;background:rgba(255,0,0,.25);padding:8px 12px;text-align:center;font-weight:900}@media(min-width: 1280px){.mobile-warning{display:none}}/*# sourceMappingURL=datagrid.css.map */
