+++ src/css/global.css
@@ -1,1 +1,1 @@
-@import"https://fonts.googleapis.com/css2?family=Comfortaa:wght@300..700&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap";*{padding:0;margin:0;font-family:"Roboto",sans-serif}h1,h2,h3,p{color:#fff;text-shadow:2px 2px 5px rgba(0,0,0,.5)}header{display:grid;align-items:center;text-align:center;height:64px}#root{display:grid;grid-template-rows:auto 1fr auto;min-height:100vh;background:#0a2647}main{display:grid;grid-auto-flow:row;align-content:space-between;justify-content:center;gap:16px;padding:32px 0px}.paginator{display:grid;justify-self:center;grid-template-columns:2fr repeat(5, 1fr) 2fr;gap:8px}.paginator .paginator-num{cursor:pointer;font-size:20px;padding:4px 10px;text-align:center;transition:.1s ease-in-out;border-radius:20px;color:#fff;justify-self:center;background:#144272}.paginator .paginator-num:hover{background:#2c74b3}.paginator .selected{background:#000}/*# sourceMappingURL=global.css.map */
+@import"https://fonts.googleapis.com/css2?family=Comfortaa:wght@300..700&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap";*{padding:0;margin:0;font-family:"Roboto",sans-serif}h1,h2,h3,p{color:#fff;text-shadow:2px 2px 5px rgba(0,0,0,.5)}header{display:grid;align-items:center;text-align:center;height:64px}footer{text-align:center;padding:8px}#root{display:grid;grid-template-rows:auto 1fr auto;min-height:100vh;background:#0a2647}main{display:grid;grid-auto-flow:row;align-content:space-between;justify-content:center;gap:16px;padding:32px 0px}.paginator{display:grid;justify-self:center;grid-template-columns:2fr repeat(5, 1fr) 2fr;gap:8px}.paginator .paginator-num{cursor:pointer;font-size:20px;padding:4px 10px;text-align:center;transition:.1s ease-in-out;border-radius:20px;color:#fff;justify-self:center;background:#144272}.paginator .paginator-num:hover{background:#2c74b3}.paginator .selected{background:#000}/*# sourceMappingURL=global.css.map */
