+++ src/main.tsx
@@ -1,11 +1,9 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-// import App from './App.tsx'
-import MasterDetail from './MasterDetail.tsx'
+import App from './App.tsx'
 
 createRoot(document.getElementById('root')!).render(
 	<StrictMode>
-		{/* <App /> */}
-		<MasterDetail/>
+		<App />
 	</StrictMode>,
 )
